In this project we analyze comments in youtube videos using NLP techniques and determine which subjects are prone to harsh comments, trolls, verbal abuse.  We will also see who are repeat offenders , correlation between number of likes/dislikes in a video vs the bullying. We then advise the YouTube admins and cyber police as precautionary measures to take.
In this project we analyze comments in youtube videos using NLP techniques and determine which subjects are prone to harsh comments, trolls, verbal abuse.  We will also see who are repeat offenders , correlation between number of likes/dislikes in a video vs the bullying. We then advise the YouTube admins and cyber police as precautionary measures to take.